test: Split memory write trigger coverage
